pyproject.toml
[project]
name = "learning-langchain-v0.3"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-core>=0.3.15",
    "langchain-openai>=0.2.5",
    "langsmith>=0.1.139",
    "langchain-community>=0.3.5",
    "langchain>=0.3.7",
    "docarray>=0.40.0",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "google-search-results>=2.4.2",
    "python-dotenv>=1.0.1",
    "langchain-google-genai>=2.0.9",
    "scikit-image>=0.25.2",
    "matplotlib>=3.10.7",
    "diffusers",
    "torch>=2.9.0",
    "google-genai>=1.46.0",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers.git" }
